services/user-service/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.17
pymysql==1.1.0
pydantic==2.0.2
python-dotenv==1.0.0
cryptography==41.0.1
email-validator==2.0.0.post2
passlib==1.7.4